spalteninhalte ueberarbeitet
